changed int to float value.

+++ app/src/main/java/com/three/tech/quickconvert/ConverterService.kt
@@ -4,7 +4,7 @@ import com.three.tech.quickconvert.networking.util.NetworkResult
 
 interface ConverterService {
 
-    suspend fun getCurrencyValue(baseCurrency: String, targetCurrency: String, amount: Int): NetworkResult
+    suspend fun getCurrencyValue(baseCurrency: String, targetCurrency: String, amount: Float): NetworkResult
 
     suspend fun doSomething()
